requirements.txt
# Kronos model core
numpy
pandas==2.2.2
torch>=2.0.0
einops==0.8.1
huggingface_hub==0.33.1
matplotlib==3.9.3
tqdm==4.67.1
safetensors==0.6.2

# Stock Advisor app
flask>=3.0
flask-cors>=4.0
akshare>=1.14
pyyaml>=6.0
